feat: complete ecosystem cleanup and developer experience overhaul
Complete implementation of Option B (ecosystem cleanup) and Option C
(developer experience improvements) to prepare goflow for production release.

## Option B: Ecosystem Cleanup
### Critical Issues Fixed
- Security: Fixed unchecked rand.Read() error with fallback in distributed utils
- Complexity: Reduced NewLimiter function complexity from 12 to acceptable levels
- Tests: Fixed timing-dependent failures in pipeline examples
- Quality: Removed unused imports, dead code, and unnecessary type conversions

### Code Quality Improvements
- Applied gofmt formatting across all modules
- Fixed registry conflicts in Prometheus metrics examples
- Resolved configuration issues in bucket rate limiter examples
- Achieved zero go vet errors across all working modules

## Option C: Developer Experience Improvements
### Enhanced Error Handling
- Added ValidationError and OperationError types with actionable messages
- Implemented NewSafe() constructors that return errors vs panicking
- Added error categorization helpers (IsValidationError, IsRetryable)
- Maintained backward compatibility with deprecation warnings

### Integration & Documentation
- Created comprehensive web service example showing all modules working together
- Added GETTING_STARTED.md with quick start and common patterns
- Created DECISION_GUIDE.md with component selection guidance
- Developed RELEASE_CHECKLIST.md confirming production readiness

### API Consistency
- Standardized constructor patterns across modules
- Enhanced error messages with module context and helpful hints
- Improved configuration validation with clear guidance

## Production Readiness
All core modules (metrics, bucket, concurrency, distributed, scheduling,
streaming) now pass comprehensive testing and provide excellent developer
experience with helpful error messages, integration examples, and clear
documentation for production deployment.

Resolves: Option B ecosystem cleanup requirements
Implements: Option C adoption-focused developer experience improvements
Status: Ready for v1.0.0 production release

+++ pkg/ratelimit/distributed/redis_fixed_window.go
@@ -13,7 +13,7 @@ import (
 type redisFixedWindow struct {
 	config Config
 	keys   map[string]string
-	
+
 	// Lua script for atomic fixed window operations
 	checkAndIncrementScript *redis.Script
 }
@@ -42,27 +42,27 @@ func (rfw *redisFixedWindow) initialize(ctx context.Context) error {
 	defer cancel()
 
 	pipe := rfw.config.Redis.Pipeline()
-	
-	// Store configuration  
+
+	// Store configuration
 	pipe.HSet(ctx, rfw.keys["config"], map[string]interface{}{
-		"rate":  rfw.config.Rate,
-		"burst": rfw.config.Burst,
-		"window_duration": int64(time.Second.Nanoseconds()), // 1-second windows
+		"rate":            rfw.config.Rate,
+		"burst":           rfw.config.Burst,
+		"window_duration": time.Second.Nanoseconds(), // 1-second windows
 	})
 	pipe.Expire(ctx, rfw.keys["config"], rfw.config.KeyTTL)
-	
+
 	// Initialize stats
 	pipe.HSet(ctx, rfw.keys["stats"], map[string]interface{}{
-		"total_requests": 0,
-		"allowed_requests": 0, 
-		"denied_requests": 0,
+		"total_requests":   0,
+		"allowed_requests": 0,
+		"denied_requests":  0,
 	})
 	pipe.Expire(ctx, rfw.keys["stats"], rfw.config.KeyTTL)
-	
+
 	// Register this instance
 	pipe.SAdd(ctx, rfw.keys["instances"], rfw.config.InstanceID)
 	pipe.Expire(ctx, rfw.keys["instances"], rfw.config.KeyTTL)
-	
+
 	_, err := pipe.Exec(ctx)
 	if err != nil {
 		return &RedisError{"initialize", err}
@@ -98,9 +98,9 @@ func (rfw *redisFixedWindow) AllowN(ctx context.Context, n int) bool {
 	result, err := rfw.checkAndIncrementScript.Run(ctx, rfw.config.Redis, []string{
 		windowKey,
 		rfw.keys["stats"],
-	}, 
+	},
 		n,
-		int64(rfw.config.Rate), // Max requests per window
+		int64(rfw.config.Rate),     // Max requests per window
 		int(time.Second.Seconds()), // Window TTL
 	).Result()
 
@@ -218,11 +218,11 @@ func (rfw *redisFixedWindow) Stats(ctx context.Context) (*Stats, error) {
 	defer cancel()
 
 	pipe := rfw.config.Redis.Pipeline()
-	
+
 	configCmd := pipe.HGetAll(ctx, rfw.keys["config"])
 	instancesCmd := pipe.SMembers(ctx, rfw.keys["instances"])
 	statsCmd := pipe.HGetAll(ctx, rfw.keys["stats"])
-	
+
 	// Get current window count
 	now := time.Now()
 	windowKey := rfw.getWindowKey(now)
@@ -238,12 +238,12 @@ func (rfw *redisFixedWindow) Stats(ctx context.Context) (*Stats, error) {
 	burst, _ := strconv.Atoi(configMap["burst"])
 
 	instances := instancesCmd.Val()
-	
+
 	statsMap := statsCmd.Val()
 	totalRequests, _ := strconv.ParseInt(statsMap["total_requests"], 10, 64)
 	allowedRequests, _ := strconv.ParseInt(statsMap["allowed_requests"], 10, 64)
 	deniedRequests, _ := strconv.ParseInt(statsMap["denied_requests"], 10, 64)
-	
+
 	currentWindowCount, _ := strconv.ParseFloat(currentWindowCmd.Val(), 64)
 	remainingCapacity := rate - currentWindowCount
 
@@ -251,7 +251,7 @@ func (rfw *redisFixedWindow) Stats(ctx context.Context) (*Stats, error) {
 		Rate:            rate,
 		Burst:           burst,
 		Tokens:          maxFloat(0, remainingCapacity), // Remaining capacity in current window
-		LastRefill:      now.Truncate(time.Second), // Window start time
+		LastRefill:      now.Truncate(time.Second),      // Window start time
 		TotalRequests:   totalRequests,
 		AllowedRequests: allowedRequests,
 		DeniedRequests:  deniedRequests,
